+++ pkg/evm/evm.go
@@ -81,7 +81,7 @@ func GetPrivateKeyBalance(
 	client ethclient.Client,
 	privateKey string,
 ) (*big.Int, error) {
-	addr, err := PrivateKeyToAddress(privateKey)
+	addr, err := utils.PrivateKeyToAddress(privateKey)
 	if err != nil {
 		return nil, err
 	}
@@ -842,12 +842,4 @@ func WaitForRPC(ctx context.Context, rpcURL string) error {
 		case <-time.After(1 * time.Second):
 		}
 	}
-}
-
-func PrivateKeyToAddress(privateKey string) (common.Address, error) {
-	pk, err := crypto.HexToECDSA(privateKey)
-	if err != nil {
-		return common.Address{}, err
-	}
-	return crypto.PubkeyToAddress(pk.PublicKey), nil
 }
